udate life expectancy test on each planet to include uranus

+++ src/solar-expectancy.js
@@ -9,12 +9,12 @@ export class SolarExpectancy {
     this.earthAgeExpectancy = this.ageExpectationService.ageExpectation(userGender, userContinent);
   }
   ageExpectancyArray() {
-    const mercuryAgeExpectancy = parseInt(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'mercury').toFixed(2));
-    const venusAgeExpectancy = parseInt(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'venus').toFixed(2));
-    const marsAgeExpectancy = parseInt(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'mars').toFixed(2));
-    const jupiterAgeExpectancy = parseInt(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'jupiter').toFixed(2));
-    const saturnAgeExpectancy = parseInt(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'saturn').toFixed(2));
-    const neptuneAgeExpectancy = parseInt(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'neptune').toFixed(2));
+    const mercuryAgeExpectancy = parseFloat(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'mercury')).toFixed(2);
+    const venusAgeExpectancy = parseFloat(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'venus')).toFixed(2);
+    const marsAgeExpectancy = parseFloat(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'mars')).toFixed(2);
+    const jupiterAgeExpectancy = parseFloat(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'jupiter')).toFixed(2);
+    const saturnAgeExpectancy = parseFloat(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'saturn')).toFixed(2);
+    const neptuneAgeExpectancy = parseFloat(this.solarConversionService.yearConversion(this.earthAgeExpectancy, 'neptune')).toFixed(2);
     return [this.earthAgeExpectancy, mercuryAgeExpectancy, venusAgeExpectancy, marsAgeExpectancy, jupiterAgeExpectancy, saturnAgeExpectancy, neptuneAgeExpectancy];
   }
 }
